[inductor] Make autotune_process.py pass mypy
TensorMeta.from_irnodes handles either a single IRNode or a tuple or
list of them, expressed here using overloading.

This changes the type of the argument to benchmark_in_sub_process to
the more specific TritonTemplateCaller, since that one has the bmreq
member and existing docstrings indicate that only the triton template
benchmark is handled.

The rand_strided call caused a mypy error because the default value
for device was a string. This is fixed by adding type hints to
rand_strided in torch/_dynamo/testing.py. Likewise, the return value
of PyCodeCache.load_by_key_path can be inferred from the type hint on
PyCodeCache.cache.

See #105230