[inductor] Enable vectorization in fbcode
Pull Request resolved: https://github.com/pytorch/pytorch/pull/105756

In fbcode, to run the test python script (with its accompanying test DSO) we
need to invoke the correct python, with the correct PYTHONPATH, so we supply
those by reading the appropriate values out of `sys`.

It's an improvement for OSS too, since the user may not be running the default
python.

My previous attempt of using `torch.backends.cpu.get_cpu_capability()` didn't work out, for two reasons:
1. That function actually refuses to report AVX512 support; it's #ifdef-ed out, for some reason.
2. In CI, we apparently are picking INVALID_VEC_ISA (at least when running
inductor_timm_cpu_accuracy), whereas `get_cpu_capability` reports AVX2.  This
is surprising, and probably indicates a bug (either in cpu capability or our
test binary), but I'd rather not go digging for it.
ghstack-source-id: 195432820

Differential Revision: [D47678649](https://our.internmc.facebook.com/intern/diff/D47678649/)