[inductor] triton_utils.config_of: check for divisibility by 16 even if expr is not an Integer
TL;DR: triton_utils.config_of determines divisibility by 16 for each of the inputs to the kernel (pointer alignment for pointers, and divisibility by 16 for sizes). For sizes, the check previously could only return true if the expr representing the size was an integer. However, it's possible for non-integral exprs to be divisible by 16, e.g. for an expr like 16*s0.

Motivation: Knowledge about divisibility by 16 allows for vectorizing loads and stores, which can improve memory bandwidth. If we have, for example, kernels with sizes [s0, 16] (dynamic batch size and static, divisible by 16 other dimensions), we want to still be able to vectorize those loads and stores.

ghstack-source-id: aeb5b9e8d644cd759e1adcb7f016f9f704f4a87a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/105743